Projet/Visualisation/requirements.txt
psycopg2-binary
ccxt
yfinance
python-dotenv
pandas
